% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_2110_cut}
\alias{data_2110_cut}
\title{IMF World Economic Outlook data, previous}
\format{
A tibble.
\describe{
\item{unit}{refer to meta unit}
\item{concept}{refer to meta concept}
\item{ref_area}{refer to meta area}
\item{freq}{"A" for annual}
\item{lastactualdate}{year like 2019}
\item{scale}{refer to meta scale}
\item{notes}{some notes}
\item{year}{year like 1980}
\item{value}{observed value}
}
}
\source{
https://www.imf.org/en/Publications/WEO/weo-database/2021/October
}
\usage{
data_2110_cut
}
\description{
Economic data published in October 2021. Cut to since 2019 only.
}
\keyword{datasets}
