% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_2404}
\alias{data_2404}
\title{IMF World Economic Outlook data, current}
\format{
A tibble.
\describe{
\item{unit}{refer to meta unit}
\item{concept}{refer to meta concept}
\item{ref_area}{refer to meta area}
\item{freq}{"A" for annual}
\item{lastactualdate}{year like 2019}
\item{scale}{refer to meta scale}
\item{notes}{some notes}
\item{year}{year like 1980}
\item{value}{observed value}
}
}
\source{
https://www.imf.org/en/Publications/WEO/weo-database/2024/April
}
\usage{
data_2404
}
\description{
Economic data published in April 2024.
}
\keyword{datasets}
