% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_chart}
\alias{draw_chart}
\title{Draw a line chart}
\usage{
draw_chart(prev, df, df_prev)
}
\arguments{
\item{prev}{"Yes" or "No".}

\item{df}{A tibble.}

\item{df_prev}{A tibble.}
}
\value{
A plotly object.
}
\description{
Draw a time-series line chart.
}
\examples{
\dontrun{
draw_chart(prev, df, df_prev)
}
}
