% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_areas}
\alias{lookup_areas}
\title{Look-up area}
\usage{
lookup_areas(code)
}
\arguments{
\item{code}{A character vector of length 1.}
}
\value{
A character vector.
}
\description{
Look-up area description from area code.
}
\examples{
\dontrun{
lookup_areas("001") # World
}
}
