% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_lastactual}
\alias{lookup_lastactual}
\title{Look-up last actual year}
\usage{
lookup_lastactual(code, df)
}
\arguments{
\item{code}{A character vector of length 1.}

\item{df}{A data frame.}
}
\value{
A year.
}
\description{
Look-up last actual year from area code.
}
\examples{
\dontrun{
lookup_lastactual("001", df)
}
}
