% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mainApp}
\alias{mainApp}
\title{Show data}
\usage{
mainApp(weo, weo_prev_data)
}
\arguments{
\item{weo}{A list.}

\item{weo_prev_data}{A data frame.}
}
\value{
A shinyApp.
}
\description{
Provide 3 panels to show data by area, region, and commodity.
}
\examples{
\dontrun{
mainApp()
}
}
