% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{menu}
\alias{menu}
\title{Menu, current}
\format{
A list.
\describe{
\item{a_menu}{
area menu

A list.
\describe{
\item{area}{a named character vector, 3 digit code}
\item{region}{a named character vector, 3 digit code}
\item{world}{a named character vector, 3 digit code, "001" named "World"}
}
}
\item{c_menu}{
concept menu

A list.
\describe{
\item{area}{a named list}
\item{region}{a named list}
\item{world}{a named list}
}
}
}
}
\source{
https://www.imf.org/en/Publications/WEO/weo-database/2022/October
}
\usage{
menu
}
\description{
Menu, current
}
\keyword{datasets}
