% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta}
\alias{meta}
\title{IMF World Economic Outlook meta data}
\format{
A list of 7 data frames
\describe{
\item{1}{frequency description}
\item{2}{date}
\item{3}{area}
\item{4}{concept}
\item{5}{denominator}
\item{6}{scale}
\item{7}{year}
}
}
\source{
https://www.imf.org/external/pubs/ft/weo/2020/01/weodata/WEOApr2020_SDMXDSD.XLSX
}
\usage{
meta
}
\description{
Meta data published in October 2020.
}
\keyword{datasets}
