% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_data}
\alias{output_data}
\title{Prepare download data}
\usage{
output_data(prev, df, df_prev, pub_ym_current, pub_ym_prev, m_area)
}
\arguments{
\item{prev}{A character vector of length 1, "Yes" or "No".}

\item{df}{A tibble.}

\item{df_prev}{A tibble.}

\item{pub_ym_current}{A character vector of length 1, like "2104".}

\item{pub_ym_prev}{A character vector of length 1, like "2010".}

\item{m_area}{A tibble, with 2 columns: code and area.}
}
\value{
A tibble.
}
\description{
Prepare download data
}
\examples{
\dontrun{
output_data(input$previous, chart_data(), chart_data_prev(),
"2104", "2010", m_area)
}
}
