% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{print_note}
\alias{print_note}
\title{Print notes}
\usage{
print_note(df)
}
\arguments{
\item{df}{A tibble.}
}
\value{
A character vector of length 1.
}
\description{
Print notes.
}
\examples{
\dontrun{
print_note(df)
}
}
