% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectArea.R
\name{selectAreaUI}
\alias{selectAreaUI}
\title{selectArea module UI}
\usage{
selectAreaUI(id, a_menu, c_menu)
}
\arguments{
\item{id}{A character vector of length 1.}

\item{a_menu}{A character vector.}

\item{c_menu}{A character vector.}
}
\value{
A module UI.
}
\description{
selectArea module UI.
}
\examples{
\dontrun{
selectAreaUI("area", area_menu, concept1_menu)
}
}
