% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weo}
\alias{weo}
\title{IMF World Economic Outlook data}
\format{
A data frame.
\describe{
\item{UNIT}{1 letter showing unit}
\item{CONCEPT}{concept}
\item{REF_AREA}{area}
\item{FREQ}{1 letter showing frequency}
\item{LASTACTUALDATE}{last actual year}
\item{SCALE}{scale}
\item{NOTES}{notes}
\item{TIME_PERIOD}{year}
\item{OBS_VALUE}{value}
}
}
\source{
https://www.imf.org/~/media/Files/Publications/WEO/WEO-Database/2020/02/weooct2020_sdmxdata.ashx
}
\usage{
weo
}
\description{
Economic data published in October 2020.
}
\keyword{datasets}
