% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weo_2104}
\alias{weo_2104}
\title{IMF World Economic Outlook data, current}
\format{
A list.
\describe{
\item{data}{

A tibble.
\describe{
\item{unit}{refer to meta unit}
\item{concept}{refer to meta concept}
\item{ref_area}{refer to meta area}
\item{freq}{"A" for annual}
\item{lastactualdate}{year like 2019}
\item{scale}{refer to meta scale}
\item{notes}{some notes}
\item{year}{year like 1980}
\item{value}{observed value}
}
}
\item{a_menu}{
area menu

A list.
\describe{
\item{area}{a named character vector, 3 digit code}
\item{region}{a named character vector, 3 digit code}
\item{world}{a named character vector, 3 digit code, "001" named "World"}
}
}
\item{c_menu}{
concept menu

A list.
\describe{
\item{area}{a named list}
\item{region}{a named list}
\item{world}{a named list}
}
}
\item{meta}{
meta data of code and description

A list.
\describe{
\item{area}{a tibble with 2 columns, code and area}
\item{concept}{a named character vector, code named by description}
\item{unit}{a named character vector, code named by description}
\item{scale}{a named character vector, code named by description}
}
}
}
}
\source{
https://www.imf.org/~/media/Files/Publications/WEO/WEO-Database/2020/02/weooct2020_sdmxdata.ashx
}
\usage{
weo_2104
}
\description{
Economic data published in October 2020.
}
\keyword{datasets}
