% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_chart}
\alias{draw_chart}
\title{Draw a line chart}
\usage{
draw_chart(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A ggplot2 plot.
}
\description{
Draw a time-series line chart.
}
\examples{
\dontrun{
draw_chart(df)
}
}
