% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mainApp}
\alias{mainApp}
\title{Show data}
\usage{
mainApp(data_current, data_prev, name_current, name_prev)
}
\arguments{
\item{data_current}{A tibble}

\item{data_prev}{A tibble.}

\item{name_current}{A character vector of length 1, like "2104".}

\item{name_prev}{A character vector of length 1, like "2010".}
}
\value{
A shinyApp.
}
\description{
Provide 3 panels to show data by area, region, and commodity.
}
\examples{
\dontrun{
mainApp()
}
}
