% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta}
\alias{meta}
\title{Meta data, current}
\format{
A list.
\describe{
\item{area}{

A tibble.
\describe{
\item{code}{"USA"}
\item{area}{"United States"}
}
}
\item{concept}{a named character vector, code named by description}
\item{unit}{a named character vector, code named by description}
\item{scale}{a named character vector, code named by description}
}
}
\source{
https://data.imf.org/en/datasets/IMF.RES:WEO
}
\usage{
meta
}
\description{
Meta data, current
}
\keyword{datasets}
