% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{print_lastactual}
\alias{print_lastactual}
\title{Print last actual year}
\usage{
print_lastactual(df)
}
\arguments{
\item{df}{A tibble.}
}
\value{
A character vector of length 1.
}
\description{
Print last actual year.
}
\examples{
\dontrun{
print_lastactual(df)
}
}
