% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectArea.R
\name{selectAreaServer}
\alias{selectAreaServer}
\title{selectArea module server}
\usage{
selectAreaServer(id, data, data_prev, name_current, name_prev)
}
\arguments{
\item{id}{A character vector of length 1.}

\item{data}{A tibble.}

\item{data_prev}{A tibble.}

\item{name_current}{A character vector of length 1, like "2104".}

\item{name_prev}{A character vector of length 1, like "2010".}
}
\value{
A module server.
}
\description{
selectArea module server
}
\examples{
\dontrun{
selectAreaServer("area", data_2110, data_2104_cut, "2110", "2104")
}
}
