% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectArea.R
\name{selectAreaServer}
\alias{selectAreaServer}
\title{selectArea module server}
\usage{
selectAreaServer(id, data, data_prev, m_area, m_concept, m_unit, m_scale)
}
\arguments{
\item{id}{A character vector of length 1.}

\item{data}{A tibble.}

\item{data_prev}{A tibble.}

\item{m_area}{A tibble with 2 columns, code and area.}

\item{m_concept}{A named character vector.}

\item{m_unit}{A named character vector.}

\item{m_scale}{A named character vector.}
}
\value{
A module server.
}
\description{
selectArea module server
}
\examples{
\dontrun{
selectAreaServer("area", weo$data, weo_prev_data,
weo$meta$area, weo$meta$concept, weo$meta$unit, weo$meta$scale)
}
}
