% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectArea.R
\name{selectAreaServer}
\alias{selectAreaServer}
\title{selectArea module server}
\usage{
selectAreaServer(id)
}
\arguments{
\item{id}{A character vector of length 1.}
}
\value{
A module server.
}
\description{
selectArea module server
}
\examples{
\dontrun{
selectAreaServer("area")
}
}
