% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weo}
\alias{weo}
\title{IMF World Economic Outlook data}
\format{
A data frame.
\describe{
\item{unit}{1 letter showing unit}
\item{concept}{concept}
\item{ref_area}{area code}
\item{freq}{1 letter showing frequency}
\item{lastactualdate}{last actual year}
\item{scale}{scale}
\item{notes}{notes}
\item{time_period}{year}
\item{obs_value}{value}
}
}
\source{
https://www.imf.org/~/media/Files/Publications/WEO/WEO-Database/2020/02/weooct2020_sdmxdata.ashx
}
\usage{
weo
}
\description{
Economic data published in October 2020.
}
\keyword{datasets}
